fix: correct agency-health error rate/uptime and requestsPerMin in analytics
F1: count failures (status <> 'success') not successes; compute errorRate as a
percentage (error_count/total*100) so uptime = 100 - errorRate is meaningful.
F3: replace integer floor-division for requestsPerMin with float round(..., 2)
so low-traffic agencies produce a non-zero rate; change schema field to float.

+++ backend/app/schemas/insight.py
@@ -39,7 +39,7 @@ class Agency(BaseModel):
     currentLatency: float
     avgLatency: float
     errorRate: float
-    requestsPerMin: int
+    requestsPerMin: float
     lastCheckedAt: datetime
 
 class Incident(BaseModel):
